requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pyjwt
bcrypt
python-multipart